requirements.txt
Flask>=3.0
markdown>=3.4
pyyaml>=6.0
pytest>=7.0
